Refactor two functions and add two functions.
Refactor durand-choose-list and durand-update-link, and add durand-mark-links
and durand-mark-weblink. I found the function durand-update-link too ugly, so I
refactor it to be more readable.

Also I refactor durand-choose-list to show the chosen items.

Finally I add a feature (durand-mark-links and durand-mark-weblink) to mark
links for actions (to mark it as seen or remove the mark or delete the link).