Fix undefined commands
There once was an eshall command called `vid`, but that is no longer there. So I
fix this and change it into the expanded, working, command.
There is another command that I have to re-define each time I call this command,
in order for me to make sure it works.