Fix some mappings and add org-agenda info on modeline
Some mappings I find not very convenient, so I fix them. Also I added a way to
show the number of tasks for today on the modeline. However, I do not want to be
bothered with such information for the moment, so I disable this functionality
again. Per chance in the future I will enable this functionality again and start
to profit from it.