Fine tune org-noter and set up link-open function.
The package 'org-noter' has some customizable options, and I set them up here.
Also for a specific PDF document, there is a problem with the property section,
so I defined an advice to remove the offending section.

In addition, I set up the function to open links in org-mode correctly: making
an argument optional is all it takes.