Add roam back.
It turns out roam is not just about the back-links. It is a method of taking
notes and a system of building a net of information.
I shall try to adapt to this workflow and see how well it goes.

I set org-roam-directory to org-directory by default, and set the template for
the file name to include separators for date strings, so that they are more
readable to me.