icomplete instead of ivy
I documented this switch in my diary.
Basically it is a transition to a simpler and more modular system.

In the mean time I made numerous changes to the settings. They are mostly due to
this transition of systems, but some of them are just minor settings I
discovered in the process of adapting to the new system, so they are also
commited together.