A bump to emacs 27
Some changes are not related to this change in the version of emacs. But some
are. For example, doom also changes the version, and a macro has been obsolete:
cmd! replaces lambda! now.

The changes that are not directly related to such changes are some attempts at
improving the performance of emacs. For example, I now use '(flex
partial-completion) instead of orderless now.