Change time related functions and add a functionality
I add an advice to org-agenda-goto-date to use evil-emacs-state after that
function.

The time related function that converts date to time now uses the results of
decode-time to feed into encode-time. This can ensure that arguments are right.

I add a function to normalize the time values so that the second, minute, and
the hour values are all 0. This can solve the problem of durand-account-report
that one has to specify a day earlier to show the desired time range.

I also add a functionality to durand-view-go-to-account-day. Now when one
executes the command conitnuously, then it will scroll the account buffer, much
like the org agenda behaviour.