Add a directory to search pdf and delete mpv thumbnails.
I forgot one place I could put pdf files in. And I now use an extension of mpv
to provide the functionality of previewing a video at a given time by
thumbnails. But these thumbnails are stored in a temporary folder, and I worry
they might not be properly deleted, so I added this cleaning functionality to
the function `clean-up-buffers'.