Adjust behaviour of bongo.
It now uses the function bongo-playlist-buffer to add the music to. This is more
robust in my opinion.

Moreover, it won't automatically play the added music. Sometimes I am listening
to other songs, and don't want to be interrupted by adding a new music.

Finally, I override the function bongo-buffer to always return
bongo-playlist-buffer. This way it won't jump to the dired buffer when it is
opened, when I launch bongo.